dpgen: move acronyms to acronyms.txt
